requirements.txt
streamlit
openai
python-dotenv
numpy
plotly
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2